build: added functionalities to prevent a user to go to a route which are not supposed to go
